refactoring logs and config endpoints
