+++ backend/api/config/application.js
@@ -1,0 +1,2 @@
+// check if debug mode
+global.debugMode = process.env.YTDL_MODE === 'debug';
