+++ backend/api/routes/logRouter.js
@@ -1,0 +1,8 @@
+
+var express = require("express");
+const { getLogs } = require("../controllers/logs");
+var logRouter = express.Router();
+
+logRouter.get('/', getLogs);
+
+module.exports = logRouter;
